nifty_3layer_system/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
yfinance>=0.1.70
loguru>=0.6.0
scipy>=1.7.0
requests>=2.28.0
python-dotenv>=0.20.0
urllib3>=1.26.0
websockets>=11.0.0
